Intial upload
Something I thought of in the shower that I though might be a viable strategy. Testing shows that it doesn't seem to be a viable trading strategy. Fails more than it succeeds. Might need to try using different tickers to see if that makes a difference. Probably need to just develop and test a different strategy.

Uploading the code in case I think of a better approach.